+++ lib/modules/datasource/gitlab-tags/util.ts
@@ -1,3 +1,4 @@
+import { GlobalConfig } from '../../../config/global';
 import { regEx } from '../../../util/regex';
 import { joinUrlParts } from '../../../util/url';
 
@@ -5,6 +6,11 @@ export const defaultRegistryUrl = 'https://gitlab.com';
 
 export function getDepHost(registryUrl: string = defaultRegistryUrl): string {
   return registryUrl.replace(regEx(/\/api\/v4$/), '');
+}
+
+export function getDefaultRegistryUrl(): string[] {
+  const { platform, endpoint } = GlobalConfig.get();
+  return platform === 'gitlab' && endpoint ? [endpoint] : [defaultRegistryUrl];
 }
 
 export function getSourceUrl(
